requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
yfinance>=0.2.30
sqlalchemy>=2.0.0
plotly>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
openai>=1.10.0
langchain-google-genai>=0.0.6
beautifulsoup4>=4.12.0
lxml>=4.9.0
kaleido>=0.2.1
fpdf2>=2.7.5
